requirements.txt
# Brain_Scape — Python Dependencies
# Pinned versions for reproducibility

# ── Neuroimaging ──
nibabel==5.2.1
nilearn==0.10.3
vtk==9.3.1
antspy==0.3.8
dipy==1.9.0
pydicom==2.4.4
mne==1.7.1

# ── Deep Learning ──
torch==2.3.1
monai==1.3.2
nnunetv2==2.5.1

# ── Serving & API ──
fastapi==0.111.0
uvicorn[standard]==0.30.1
celery[redis]==5.4.0
redis==5.0.7
python-multipart==0.0.9
websockets==12.0

# ── MLOps ──
prefect==2.19.0
mlflow==2.14.1

# ── Compliance & Security ──
presidio-analyzer==2.2.35
presidio-anonymizer==2.2.35
cryptography==42.0.8
PyJWT==2.8.0

# ── LLM & RAG ──
langchain==0.2.6
langchain-openai==0.1.10
pinecone-client==4.1.1
weaviate-client==4.6.1
openai==1.35.3
sentence-transformers==3.0.1

# ── OCR ──
pytesseract==0.3.10
Pillow==10.3.0

# ── Report Generation ──
reportlab==4.2.0
weasyprint==62.3
jinja2==3.1.4

# ── Data & DB ──
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2
boto3==1.34.138

# ── Monitoring ──
prometheus-client==0.20.0

# ── Voice ──
openai-whisper==20231117

# ── Testing ──
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
pytest-cov==5.0.0

# ── Utilities ──
pyyaml==6.0.1
python-dotenv==1.0.1
click==8.1.7
rich==13.7.1
numpy==1.26.4
scipy==1.14.0